% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covidcases}
\alias{covidcases}
\title{Reported Covid-19 infections Germany}
\format{
A daily time series of the class data.frame.
}
\source{
Robert Koch-Institut (RKI), \href{https://www.rki.de/DE/Content/InfAZ/N/Neuartiges_Coronavirus/Daten/Fallzahlen_Kum_Tab_aktuell}{Source}, \href{https://www.govdata.de/dl-de/by-2-0}{DL-DE BY-2.0}
}
\usage{
data(covidcases)
}
\description{
Reported Covid-19 infections Germany
}
\keyword{datasets}
