% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covidhosp}
\alias{covidhosp}
\title{Covid-19 hospitalization incidence number Germany}
\format{
A daily maulivariate time series of the class data.frame.
}
\source{
Robert Koch-Institut (RKI), \href{https://www.rki.de/DE/Content/InfAZ/N/Neuartiges_Coronavirus/Daten/Fallzahlen_Kum_Tab_aktuell}{Source}, \href{https://www.govdata.de/dl-de/by-2-0}{DL-DE BY-2.0}
}
\usage{
data(covidhosp)
}
\description{
The Covid-19 hospitalization incidence is the number of hospitalized infected persons in the past 7 days per 100000 inhabitants.
The data are available for the individual federal states and for Germany overall.
}
\keyword{datasets}
