% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdpgr}
\alias{gdpgr}
\title{German Quarterly Gross Domestic Product Annual Growth Rate (nominal)}
\format{
A quarterly time series of the class ts.
}
\source{
Statistisches Bundesamt (Destatis), Genesis-Online, Tablename 81000BV007, License \href{https://www.govdata.de/dl-de/by-2-0}{DL-DE BY-2.0}
}
\usage{
data(gdpgr)
}
\description{
German Quarterly Gross Domestic Product Annual Growth Rate (nominal)
}
\keyword{datasets}
