% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{infl}
\alias{infl}
\title{Inflation Rate Germany}
\format{
A monthly time series of the class ts.
}
\source{
Statistisches Bundesamt (Destatis), Genesis-Online, Tablename 61111BM001, License \href{https://www.govdata.de/dl-de/by-2-0}{DL-DE BY-2.0}
}
\usage{
data(infl)
}
\description{
Monthly CPI inflation rate (year-on-year change). Calculated from the data set cpi as diff(log(cpi)*100,12)
}
\keyword{datasets}
