% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{temp}
\alias{temp}
\title{Average Temperature Germany}
\format{
A yearly time series of the class ts.
}
\source{
Deutscher Wetterdienst (DWD), \href{https://www.dwd.de/DE/leistungen/zeitreihen/zeitreihen.htm}{Source}, \href{https://www.dwd.de/DE/service/copyright/copyright_node.html}{Copyright}
}
\usage{
data(temp)
}
\description{
Average Temperature Germany
}
\keyword{datasets}
