% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{urate}
\alias{urate}
\title{Unemployment rate Germany}
\format{
A monthly time series of the class ts.
}
\source{
tatistisches Bundesamt (Destatis), Genesis-Online, Tablename 13211BM001, License \href{https://www.govdata.de/dl-de/by-2-0}{DL-DE BY-2.0}
}
\usage{
data(urate)
}
\description{
Monthly unemployment rates for Germany
}
\keyword{datasets}
